[toolbox] pimp up terminal type
This is a tricky one, starting with zsh-syntax-highlighting not showing
suggestions in light gray but rather the regular white color. It appears
the light gray is actually `fg=8` in the xterm-256color chart.

https://github.com/zsh-users/zsh-autosuggestions/tree/a411ef3#suggestion-highlight-style

Docker sets the TERM environment variable to `xterm` by default since as
far back as 2016 https://askubuntu.com/a/736498...

https://docs.docker.com/engine/reference/run/#env-environment-variables

It is possible to override it with the ENV directive in the dockerfile,
but it cannot be as granular as whether the container is tty-enabled or
not. So better set it in the .zshrc file.

By the way, it has the nice side-effect of fixing the fzf preview
window, which allows us to enable it in fzf-tab.

https://github.com/junegunn/fzf/tree/9499910#preview-window

Not with `tldr` though, it is still broken...

    $ tldr --update
    tldr: /root/.local/share/tldr: createDirectory: does not exist (No such file or directory)