[linux-headers] set pin priority on testing
With the default priority, testing packages are prioritized over stable.
This is not the expected behavior, and instead the target release should
be explicitly set when installing from testing. This is achieved using :

- https://serverfault.com/a/382101
- https://manpages.debian.org/bullseye/apt/apt_preferences.5.en.html

This is not using the default release https://unix.stackexchange.com/a/647205.

Note that printf is favored over echo https://github.com/koalaman/shellcheck/wiki/SC2028.

This was discovered upon running `pip install poetry` as it triggered
https://peps.python.org/pep-0668/. See also https://stackoverflow.com/a/75722775.